requirements.txt
atlassian-python-api>=3.41.2
PyYAML>=6.0.1
typing-extensions>=4.8.0
